Create reusable EmptyState, use in NoActiveTickets
